xml-hddm.cpp, hddm-xml.cpp, hddm-c.cpp, hddm-cpp.cpp [rtj]    - replaced declarations of type "long long int" with "int64_t".  The      long long declaration refers to 64-bit integers on 32-bit architectures,      but on 64-bit architectures it does not.  The int64_t is less ambiguous.

+++ programs/Analysis/hddm/hddm-xml.cpp
@@ -322,13 +322,13 @@ void XMLmaker::constructXML(xstream::xdr::istream& ifx,
          std::ostringstream attrStr;
          if (typeS == "int")
          {
-            int value;
+            int32_t value;
 	    ifx >> value;
             attrStr << " " << nameS << "=\"" << value << "\"";
          }
 	 else if (typeS == "long")
          {
-            long long value;
+            int64_t value;
             ifx >> value;
             attrStr << " " << nameS << "=\"" << value << "\"";
          }
